Fixed wrong usage of actions/checkout
Since actions/checkout@v2 the repository gets checked out with a depth of 1.
This results in a single commit containing the entire code,
which makes it imposible to rebase my fork on top of upstream.
Fixed by setting the depth to 0 (disables depth limit)

Also switched to environmental variables and removed unneeded code
to reduce the script complexity.